+++ lib/src/features/configuration/data/titles_provider.dart
@@ -1,3 +1,4 @@
+import 'package:catechism/src/features/configuration/domain/localized_titles.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-Provider<List<String>>? titlesProvider;
+Provider<LocalizedTitles>? titlesProvider;