+++ lib/src/routing/app_router.dart
@@ -1,4 +1,4 @@
-import 'package:catechism/src/features/custom_pages/domain/custom_page_data.dart';
+import 'package:catechism/src/features/custom_pages/domain/localized_custom_pages.dart';
 import 'package:catechism/src/features/custom_pages/presentation/custom_page_screen.dart';
 import 'package:catechism/src/features/not_found/not_found_screen.dart';
 import 'package:catechism/src/features/questions/presentation/screens/question_details_screen.dart';
@@ -12,7 +12,7 @@ import 'package:go_router/go_router.dart';
 
 class CatechismRouter {
   /// The customPages property is used to create the list of info screens.
-  final List<CustomPageData> customPages;
+  final LocalizedCustomPages customPages;
 
   /// The CatechismRouter constructor is used to create a catechism router.
   CatechismRouter({
@@ -30,14 +30,14 @@ class CatechismRouter {
           name: AppRoute.home.name,
           builder: (context, state) => QuestionListScreen(),
           routes: [
-            ...customPages.map(
-              (infoScreen) => GoRoute(
-                path: infoScreen.routeName,
-                name: infoScreen.routeName,
-                builder: (context, state) =>
-                    CustomPageScreen(infoScreenData: infoScreen),
-              ),
-            ),
+            ...customPages.getAllCustomPages().map(
+                  (infoScreen) => GoRoute(
+                    path: infoScreen.routeName,
+                    name: infoScreen.routeName,
+                    builder: (context, state) =>
+                        CustomPageScreen(infoScreenData: infoScreen),
+                  ),
+                ),
             GoRoute(
               path: 'details/:questionId',
               name: AppRoute.details.name,